+++ changelog.rst
@@ -2,6 +2,12 @@ ChangeLog
 =========
 
 
+`6.0.3 (unreleased) <https://github.com/kdeldycke/mail-deduplicate/compare/v6.0.2...develop>`_
+----------------------------------------------------------------------------------------------
+
+.. note:: This version is not yet released and is under active development
+
+
 `6.0.2 (2020-11-05) <https://github.com/kdeldycke/mail-deduplicate/compare/v6.0.1...v6.0.2>`_
 ---------------------------------------------------------------------------------------------